@@ -1,6 +1,6 @@
 {
   "name": "discord.js",
-  "version": "14.2.0",
+  "version": "14.3.0",
   "description": "A powerful library for interacting with the Discord API",
   "scripts": {
     "test": "yarn docs:test && yarn test:typescript",
@@ -54,19 +54,19 @@
     "@discordjs/rest": "workspace:^",
     "@sapphire/snowflake": "^3.2.2",
     "@types/ws": "^8.5.3",
-    "discord-api-types": "^0.36.3",
+    "discord-api-types": "^0.37.4",
     "fast-deep-equal": "^3.1.3",
     "lodash.snakecase": "^4.1.1",
     "tslib": "^2.4.0",
-    "undici": "^5.8.2",
+    "undici": "^5.9.1",
     "ws": "^8.8.1"
   },
   "devDependencies": {
     "@discordjs/docgen": "workspace:^",
-    "@favware/cliff-jumper": "^1.8.6",
-    "@types/node": "^16.11.47",
+    "@favware/cliff-jumper": "^1.8.7",
+    "@types/node": "^16.11.54",
     "dtslint": "^4.2.1",
-    "eslint": "^8.21.0",
+    "eslint": "^8.22.0",
     "jest": "^28.1.3",
     "prettier": "^2.7.1",
     "tsd": "^0.22.0",
