@@ -1,17 +1,17 @@
 {
 	"name": "@discordjs/docgen",
-	"version": "0.12.0",
+	"version": "0.12.1",
 	"description": "The docs.json generator for discord.js and its related projects",
 	"scripts": {
 		"build": "unbuild",
-		"lint": "prettier --check . && eslint src --ext mjs,js,ts",
-		"format": "prettier --write . && eslint src --ext mjs,js,ts --fix",
+		"lint": "prettier --check . && TIMING=1 eslint src --ext mjs,js,ts",
+		"format": "prettier --write . && TIMING=1 eslint src --ext mjs,js,ts --fix",
 		"fmt": "yarn format",
-		"prepack": "yarn lint && yarn test && yarn build",
+		"prepack": "yarn format && yarn build",
 		"changelog": "git cliff --prepend ./CHANGELOG.md -u -c ./cliff.toml -r ../../ --include-path 'packages/docgen/*'",
 		"release": "cliff-jumper"
 	},
-	"bin": "./dist/index.cjs",
+	"bin": "./dist/cli.cjs",
 	"directories": {
 		"lib": "src"
 	},
@@ -42,23 +42,31 @@
 	},
 	"homepage": "https://discord.js.org",
 	"dependencies": {
-		"@discordjs/collection": "workspace:^",
 		"commander": "^9.4.0",
 		"jsdoc-to-markdown": "^7.1.1",
 		"tslib": "^2.4.0",
 		"typedoc": "^0.23.10"
 	},
 	"devDependencies": {
-		"@favware/cliff-jumper": "^1.8.6",
+		"@favware/cliff-jumper": "^1.8.7",
 		"@types/jsdoc-to-markdown": "^7.0.3",
-		"@types/node": "^16.11.47",
-		"eslint": "^8.21.0",
+		"@types/node": "^16.11.54",
+		"@typescript-eslint/eslint-plugin": "^5.34.0",
+		"@typescript-eslint/parser": "^5.34.0",
+		"eslint": "^8.22.0",
+		"eslint-config-marine": "^9.4.1",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-import-resolver-typescript": "^3.5.0",
+		"eslint-plugin-import": "^2.26.0",
 		"prettier": "^2.7.1",
-		"rollup-plugin-typescript2": "0.32.1",
+		"rollup-plugin-typescript2": "^0.33.0",
 		"typescript": "^4.7.4",
-		"unbuild": "^0.8.4"
+		"unbuild": "^0.8.9"
 	},
 	"engines": {
 		"node": ">=16.9.0"
+	},
+	"publishConfig": {
+		"access": "public"
 	}
 }
