@@ -4,8 +4,8 @@
 	"description": "Lightweight HTTP proxy for Discord's API, brought to you as a container 📦",
 	"scripts": {
 		"build": "unbuild",
-		"lint": "prettier --check . && eslint src --ext mjs,js,ts",
-		"format": "prettier --write . && eslint src --ext mjs,js,ts --fix",
+		"lint": "prettier --check . && TIMING=1 eslint src --ext mjs,js,ts",
+		"format": "prettier --write . && TIMING=1 eslint src --ext mjs,js,ts --fix",
 		"fmt": "yarn format",
 		"prepack": "yarn lint && yarn test && yarn build",
 		"changelog": "git cliff --prepend ./CHANGELOG.md -u -c ./cliff.toml -r ../../ --include-path 'packages/proxy-container/*'"
@@ -44,17 +44,23 @@
 	},
 	"homepage": "https://discord.js.org",
 	"dependencies": {
-		"@discordjs/proxy": "^1.0.0",
-		"@discordjs/rest": "^1.0.0",
+		"@discordjs/proxy": "^1.1.0",
+		"@discordjs/rest": "^1.1.0",
 		"tslib": "^2.4.0"
 	},
 	"devDependencies": {
-		"@types/node": "^16.11.47",
-		"eslint": "^8.21.0",
+		"@types/node": "^16.11.54",
+		"@typescript-eslint/eslint-plugin": "^5.34.0",
+		"@typescript-eslint/parser": "^5.34.0",
+		"eslint": "^8.22.0",
+		"eslint-config-marine": "^9.4.1",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-import-resolver-typescript": "^3.5.0",
+		"eslint-plugin-import": "^2.26.0",
 		"prettier": "^2.7.1",
-		"rollup-plugin-typescript2": "0.32.1",
+		"rollup-plugin-typescript2": "^0.33.0",
 		"typescript": "^4.7.4",
-		"unbuild": "^0.8.4"
+		"unbuild": "^0.8.9"
 	},
 	"engines": {
 		"node": ">=16.9.0"
