@@ -1,14 +1,14 @@
 {
 	"name": "@discordjs/rest",
-	"version": "1.0.1",
+	"version": "1.1.0",
 	"description": "The REST API for discord.js",
 	"scripts": {
 		"test": "vitest run",
 		"build": "unbuild",
-		"lint": "prettier --check . && eslint src __tests__ --ext mjs,js,ts",
-		"format": "prettier --write . && eslint src __tests__ --ext mjs,js,ts --fix",
+		"lint": "prettier --check . && TIMING=1 eslint src __tests__ --ext mjs,js,ts",
+		"format": "prettier --write . && TIMING=1 eslint src __tests__ --ext mjs,js,ts --fix",
 		"fmt": "yarn format",
-		"docs": "downlevel-dts . docs --to=3.7 && docgen -i src/index.ts -c docs/index.json -o docs/docs.json --typescript && api-extractor run --local",
+		"docs": "downlevel-dts dist docs/dist --to=3.7 && docgen -i src/index.ts -c docs/index.json -o docs/docs.json --typescript && api-extractor run --local",
 		"prepack": "yarn lint && yarn test && yarn build",
 		"changelog": "git cliff --prepend ./CHANGELOG.md -u -c ./cliff.toml -r ../../ --include-path 'packages/rest/*'",
 		"release": "cliff-jumper"
@@ -53,26 +53,33 @@
 	"homepage": "https://discord.js.org",
 	"dependencies": {
 		"@discordjs/collection": "workspace:^",
-		"@sapphire/async-queue": "^1.4.0",
+		"@sapphire/async-queue": "^1.5.0",
 		"@sapphire/snowflake": "^3.2.2",
-		"discord-api-types": "^0.36.3",
+		"discord-api-types": "^0.37.4",
 		"file-type": "^17.1.6",
 		"tslib": "^2.4.0",
-		"undici": "^5.8.2"
+		"undici": "^5.9.1"
 	},
 	"devDependencies": {
 		"@discordjs/docgen": "workspace:^",
-		"@favware/cliff-jumper": "^1.8.6",
-		"@microsoft/api-extractor": "^7.29.2",
-		"@types/node": "^16.11.47",
-		"c8": "^7.12.0",
+		"@favware/cliff-jumper": "^1.8.7",
+		"@microsoft/api-extractor": "^7.29.3",
+		"@types/node": "^16.11.54",
+		"@typescript-eslint/eslint-plugin": "^5.34.0",
+		"@typescript-eslint/parser": "^5.34.0",
+		"@vitest/coverage-c8": "^0.22.1",
 		"downlevel-dts": "^0.10.0",
-		"eslint": "^8.21.0",
+		"eslint": "^8.22.0",
+		"eslint-config-marine": "^9.4.1",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-import-resolver-typescript": "^3.5.0",
+		"eslint-plugin-import": "^2.26.0",
+		"eslint-plugin-tsdoc": "^0.2.16",
 		"prettier": "^2.7.1",
-		"rollup-plugin-typescript2": "0.32.1",
+		"rollup-plugin-typescript2": "^0.33.0",
 		"typescript": "^4.7.4",
-		"unbuild": "^0.8.4",
-		"vitest": "^0.21.1"
+		"unbuild": "^0.8.9",
+		"vitest": "^0.22.1"
 	},
 	"engines": {
 		"node": ">=16.9.0"
