@@ -5,10 +5,10 @@
 	"scripts": {
 		"build": "unbuild",
 		"test": "jest --coverage",
-		"lint": "prettier --check . && eslint src __tests__ --ext mjs,js,ts",
-		"format": "prettier --write . && eslint src __tests__ --ext mjs,js,ts --fix",
+		"lint": "prettier --check . && TIMING=1 eslint src __tests__ --ext mjs,js,ts",
+		"format": "prettier --write . && TIMING=1 eslint src __tests__ --ext mjs,js,ts --fix",
 		"fmt": "yarn format",
-		"docs": "downlevel-dts . docs --to=3.7 && docgen -i src/index.ts -c docs/index.json -o docs/docs.json --typescript && api-extractor run --local",
+		"docs": "downlevel-dts dist docs/dist --to=3.7 && docgen -i src/index.ts -c docs/index.json -o docs/docs.json --typescript && api-extractor run --local",
 		"prepack": "yarn lint && yarn test && yarn build",
 		"changelog": "git cliff --prepend ./CHANGELOG.md -u -c ./cliff.toml -r ../../ --include-path 'packages/voice/*'",
 		"release": "cliff-jumper"
@@ -53,31 +53,37 @@
 	"homepage": "https://discord.js.org",
 	"dependencies": {
 		"@types/ws": "^8.5.3",
-		"discord-api-types": "^0.36.3",
+		"discord-api-types": "^0.37.4",
 		"prism-media": "^1.3.4",
 		"tslib": "^2.4.0",
 		"ws": "^8.8.1"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.18.10",
+		"@babel/core": "^7.18.13",
 		"@babel/preset-env": "^7.18.10",
 		"@babel/preset-typescript": "^7.18.6",
 		"@discordjs/docgen": "workspace:^",
-		"@discordjs/scripts": "workspace:^",
-		"@favware/cliff-jumper": "^1.8.6",
-		"@microsoft/api-extractor": "^7.29.2",
-		"@types/jest": "^28.1.6",
-		"@types/node": "^16.11.47",
+		"@favware/cliff-jumper": "^1.8.7",
+		"@microsoft/api-extractor": "^7.29.3",
+		"@types/jest": "^28.1.7",
+		"@types/node": "^16.11.54",
+		"@typescript-eslint/eslint-plugin": "^5.34.0",
+		"@typescript-eslint/parser": "^5.34.0",
 		"downlevel-dts": "^0.10.0",
-		"eslint": "^8.21.0",
+		"eslint": "^8.22.0",
+		"eslint-config-marine": "^9.4.1",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-import-resolver-typescript": "^3.5.0",
+		"eslint-plugin-import": "^2.26.0",
+		"eslint-plugin-tsdoc": "^0.2.16",
 		"jest": "^28.1.3",
 		"jest-websocket-mock": "^2.4.0",
 		"mock-socket": "^9.1.5",
 		"prettier": "^2.7.1",
-		"rollup-plugin-typescript2": "0.32.1",
+		"rollup-plugin-typescript2": "^0.33.0",
 		"tweetnacl": "^1.0.3",
 		"typescript": "^4.7.4",
-		"unbuild": "^0.8.4"
+		"unbuild": "^0.8.9"
 	},
 	"engines": {
 		"node": ">=16.9.0"
