@@ -1,7 +1,7 @@
 /**
  * @type {import('next').NextConfig}
  */
-module.exports = {
+export default {
 	reactStrictMode: true,
 	swcMinify: true,
 	eslint: {
