@@ -1,4 +1,5 @@
 import type { APIMessageComponentEmoji, APISelectMenuOption } from 'discord-api-types/v10';
+import type { JSONEncodable } from '../../util/jsonEncodable';
 
 import {
 	defaultValidator,
@@ -10,7 +11,7 @@ import {
 /**
  * Represents a option within a select menu component
  */
-export class SelectMenuOptionBuilder {
+export class SelectMenuOptionBuilder implements JSONEncodable<APISelectMenuOption> {
 	public constructor(public data: Partial<APISelectMenuOption> = {}) {}
 
 	/**
@@ -63,6 +64,9 @@ export class SelectMenuOptionBuilder {
 		return this;
 	}
 
+	/**
+	 * {@inheritDoc ComponentBuilder.toJSON}
+	 */
 	public toJSON(): APISelectMenuOption {
 		validateRequiredSelectMenuOptionParameters(this.data.label, this.data.value);
 		// eslint-disable-next-line @typescript-eslint/consistent-type-assertions
