@@ -1,3 +1,9 @@
 import { createUnbuildConfig } from '../../build.config';
 
-export default createUnbuildConfig({ minify: true, externals: ['package.cjs', 'package.mjs'] });
+export default createUnbuildConfig({
+	entries: [
+		{ builder: 'rollup', input: 'src/index' },
+		{ builder: 'rollup', input: 'src/cli' },
+	],
+	minify: true,
+});
