@@ -1,7 +1,7 @@
 import type { Type } from './index.js';
 
 export interface Return {
-	type: Type;
+	type: Required<Type>;
 	nullable?: boolean;
 	description?: string;
 }
