@@ -1,7 +1,7 @@
 import type { Type } from './index.js';
 
 export interface VarType extends Type {
-	type?: Required<Type>;
-	description?: string;
-	nullable?: boolean;
+	type?: Required<Type> | undefined;
+	description?: string | undefined;
+	nullable?: boolean | undefined;
 }
