@@ -13,6 +13,6 @@ export class AudioPlayerError extends Error {
 		super(error.message);
 		this.resource = resource;
 		this.name = error.name;
-		this.stack = error.stack;
+		this.stack = error.stack!;
 	}
 }
