@@ -4,13 +4,13 @@ import { EventEmitter } from 'node:events';
 export interface SpeakingMap extends EventEmitter {
 	/**
 	 * Emitted when a user starts speaking.
-	 * @event
+	 * @eventProperty
 	 */
 	on(event: 'start', listener: (userId: string) => void): this;
 
 	/**
 	 * Emitted when a user ends speaking.
-	 * @event
+	 * @eventProperty
 	 */
 	on(event: 'end', listener: (userId: string) => void): this;
 }
