@@ -1,24 +1,31 @@
-import type { ApiItem, ApiModel, ApiTypeParameterListMixin } from '@microsoft/api-extractor-model';
-import type { DocItemConstructor } from './DocItem';
-import { generateTypeParamData, TypeParameterData } from '~/util/parse.server';
+import type { ApiItem, ApiModel, TypeParameter } from '@microsoft/api-extractor-model';
+import { block, type DocBlockJSON } from './comment/CommentBlock';
+import { genToken, type TokenDocumentation } from '~/util/parse.server';
 
-export function TypeParameterMixin<TBase extends DocItemConstructor>(Base: TBase) {
-	return class Mixed extends Base {
-		public readonly typeParameters: TypeParameterData[] = [];
+export interface TypeParameterData {
+	name: string;
+	constraintTokens: TokenDocumentation[];
+	defaultTokens: TokenDocumentation[];
+	optional: boolean;
+	commentBlock: DocBlockJSON | null;
+}
 
-		public constructor(...args: any[]);
-		public constructor(model: ApiModel, item: ApiItem) {
-			super(model, item);
-			this.typeParameters = (item as ApiTypeParameterListMixin).typeParameters.map((typeParam) =>
-				generateTypeParamData(this.model, typeParam),
-			);
-		}
+export function generateTypeParamData(
+	model: ApiModel,
+	typeParam: TypeParameter,
+	version: string,
+	parentItem?: ApiItem,
+): TypeParameterData {
+	const constraintTokens = typeParam.constraintExcerpt.spannedTokens.map((token) => genToken(model, token, version));
+	const defaultTokens = typeParam.defaultTypeExcerpt.spannedTokens.map((token) => genToken(model, token, version));
 
-		public override toJSON() {
-			return {
-				...super.toJSON(),
-				typeParameterData: this.typeParameters,
-			};
-		}
+	return {
+		name: typeParam.name,
+		constraintTokens,
+		defaultTokens,
+		optional: typeParam.isOptional,
+		commentBlock: typeParam.tsdocTypeParamBlock
+			? block(typeParam.tsdocTypeParamBlock, model, version, parentItem)
+			: null,
 	};
 }
