@@ -1,18 +1,27 @@
 import type { ApiItem, ApiModel } from '@microsoft/api-extractor-model';
 import type { DocDeclarationReference, DocLinkTag } from '@microsoft/tsdoc';
-import { CommentNode } from './CommentNode';
+import { type DocNodeJSON, node } from './CommentNode';
 import { generatePath, resolveName } from '~/util/parse.server';
+
+interface LinkTagCodeLink {
+	name: string;
+	kind: string;
+	path: string;
+}
+
+export interface DocLinkTagJSON extends DocNodeJSON {
+	text: string | null;
+	codeDestination: LinkTagCodeLink | null;
+	urlDestination: string | null;
+}
 
 export function genToken(
 	model: ApiModel,
 	ref: DocDeclarationReference,
 	context: ApiItem | null,
+	version: string,
 ): LinkTagCodeLink | null {
-	if (!context) {
-		return null;
-	}
-
-	const item = model.resolveDeclarationReference(ref, context).resolvedApiItem ?? null;
+	const item = model.resolveDeclarationReference(ref, context ?? undefined).resolvedApiItem ?? null;
 
 	if (!item) {
 		return null;
@@ -21,34 +30,31 @@ export function genToken(
 	return {
 		name: resolveName(item),
 		kind: item.kind,
-		path: generatePath(item.getHierarchy()),
+		path: generatePath(item.getHierarchy(), version),
 	};
 }
 
-export interface LinkTagCodeLink {
-	name: string;
-	kind: string;
-	path: string;
-}
+export function linkTagNode(
+	linkNode: DocLinkTag,
+	model: ApiModel,
+	version: string,
+	parentItem?: ApiItem,
+): DocLinkTagJSON {
+	// If we weren't provided a parent object, fallback to the package entrypoint.
+	const packageEntryPoint = linkNode.codeDestination?.importPath
+		? model.getAssociatedPackage()?.findEntryPointsByPath(linkNode.codeDestination.importPath)[0]
+		: null;
 
-export class LinkTagCommentNode extends CommentNode<DocLinkTag> {
-	public readonly codeDestination: LinkTagCodeLink | null;
-	public readonly text: string | null;
-	public readonly urlDestination: string | null;
+	const codeDestination = linkNode.codeDestination
+		? genToken(model, linkNode.codeDestination, parentItem ?? packageEntryPoint ?? null, version)
+		: null;
+	const text = linkNode.linkText ?? null;
+	const urlDestination = linkNode.urlDestination ?? null;
 
-	public constructor(node: DocLinkTag, model: ApiModel, parentItem?: ApiItem) {
-		super(node, model, parentItem);
-		this.codeDestination = node.codeDestination ? genToken(model, node.codeDestination, this.parentItem) : null;
-		this.text = node.linkText ?? null;
-		this.urlDestination = node.urlDestination ?? null;
-	}
-
-	public override toJSON() {
-		return {
-			...super.toJSON(),
-			text: this.text,
-			codeDestination: this.codeDestination,
-			urlDestination: this.urlDestination,
-		};
-	}
+	return {
+		...node(linkNode),
+		text,
+		codeDestination,
+		urlDestination,
+	};
 }
