@@ -1,31 +1,27 @@
+import { Box, ScrollArea } from '@mantine/core';
 import { HyperlinkedText } from './HyperlinkedText';
 import { Table } from './Table';
+import { TSDoc } from './tsdoc/TSDoc';
 import type { ParameterDocumentation } from '~/util/parse.server';
 
-interface ParameterDetailProps {
-	data: ParameterDocumentation[];
-	className?: string | undefined;
-}
+const columnStyles = {
+	Name: 'font-mono whitespace-nowrap',
+	Type: 'font-mono whitespace-pre-wrap break-normal',
+};
 
-export function ParameterTable({ data, className }: ParameterDetailProps) {
+export function ParameterTable({ data }: { data: ParameterDocumentation[] }) {
 	const rows = data.map((param) => ({
 		Name: param.name,
 		Type: <HyperlinkedText tokens={param.tokens} />,
 		Optional: param.isOptional ? 'Yes' : 'No',
-		Description: 'None',
+		Description: param.paramCommentBlock ? <TSDoc node={param.paramCommentBlock} /> : 'None',
 	}));
 
-	const columnStyles = {
-		Name: 'font-mono',
-		Type: 'font-mono',
-	};
-
 	return (
-		<Table
-			className={className}
-			columns={['Name', 'Type', 'Optional', 'Description']}
-			rows={rows}
-			columnStyles={columnStyles}
-		/>
+		<Box>
+			<ScrollArea pb="xs" offsetScrollbars>
+				<Table columns={['Name', 'Type', 'Optional', 'Description']} rows={rows} columnStyles={columnStyles} />
+			</ScrollArea>
+		</Box>
 	);
 }
