@@ -1,16 +1,23 @@
+import { Stack } from '@mantine/core';
 import { CodeListing } from './CodeListing';
-import type { DocProperty } from '~/DocModel/DocProperty';
+import type { ApiPropertyItemJSON } from '~/DocModel/ApiNodeJSONEncoder';
 
-export interface PropertyListProps {
-	data: ReturnType<DocProperty['toJSON']>[];
-}
-
-export function PropertyList({ data }: PropertyListProps) {
+export function PropertyList({ data }: { data: ApiPropertyItemJSON[] }) {
 	return (
-		<div className="flex flex-col gap-5">
+		<Stack>
 			{data.map((prop) => (
-				<CodeListing key={prop.name} name={prop.name} typeTokens={prop.propertyTypeTokens} summary={prop.summary} />
+				<CodeListing
+					key={prop.name}
+					name={prop.name}
+					typeTokens={prop.propertyTypeTokens}
+					readonly={prop.readonly}
+					optional={prop.optional}
+					summary={prop.summary}
+					comment={prop.comment}
+					deprecation={prop.deprecated}
+					inheritanceData={prop.inheritanceData}
+				/>
 			))}
-		</div>
+		</Stack>
 	);
 }
