@@ -1,10 +1,6 @@
 import { DocContainer } from '../DocContainer';
-import type { DocVariable } from '~/DocModel/DocVariable';
+import type { ApiVariableJSON } from '~/DocModel/ApiNodeJSONEncoder';
 
-export interface VariableProps {
-	data: ReturnType<DocVariable['toJSON']>;
-}
-
-export function Variable({ data }: VariableProps) {
+export function Variable({ data }: { data: ApiVariableJSON }) {
 	return <DocContainer name={data.name} kind={data.kind} excerpt={data.excerpt} summary={data.summary} />;
 }
