@@ -2,9 +2,9 @@ import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
 
 export default function middleware(request: NextRequest) {
-	return NextResponse.rewrite(new URL('/docs/main/packages/builders', request.url));
+	return NextResponse.redirect(new URL('/docs/packages', request.url));
 }
 
 export const config = {
-	matcher: ['/docs', '/docs/:branch', '/docs/:branch/packages'],
+	matcher: ['/docs'],
 };
