@@ -1,24 +1,30 @@
-import { Html, Head, Main, NextScript } from 'next/document';
+import { createGetInitialProps } from '@mantine/next';
+import Document, { Html, Head, Main, NextScript } from 'next/document';
 
-export default function Document() {
-	return (
-		<Html lang="en">
-			<Head />
-			<body>
-				<script
-					dangerouslySetInnerHTML={{
-						__html: `(() => {
-							const prefersDarkMode = window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches;
-							const persistedColorPreference = localStorage.getItem('theme') || 'auto';
-							if (persistedColorPreference === 'dark' || (prefersDarkMode && persistedColorPreference !== 'light')) {
-								document.documentElement.classList.toggle('dark', true);
-							}
-						})();`,
-					}}
-				/>
-				<Main />
-				<NextScript />
-			</body>
-		</Html>
-	);
+const getInitialProps = createGetInitialProps();
+
+export default class _Document extends Document {
+	public static override getInitialProps = getInitialProps;
+
+	public override render() {
+		return (
+			<Html lang="en">
+				<Head>
+					<link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
+					<link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
+					<link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
+					<link rel="manifest" href="/site.webmanifest" />
+					<link rel="mask-icon" href="/safari-pinned-tab.svg" color="#090a16" />
+					<meta name="apple-mobile-web-app-title" content="discord.js" />
+					<meta name="application-name" content="discord.js" />
+					<meta name="msapplication-TileColor" content="#090a16" />
+					<meta name="theme-color" content="#1a1b1e" />
+				</Head>
+				<body>
+					<Main />
+					<NextScript />
+				</body>
+			</Html>
+		);
+	}
 }
