@@ -1,73 +1,133 @@
+import { createStyles, Container, Title, Button, Group, Text, Center, Box, useMantineColorScheme } from '@mantine/core';
 import Image from 'next/future/image';
 import Link from 'next/link';
-import { forwardRef, MouseEventHandler, Ref } from 'react';
-import codeSample from '../assets/code-sample.png';
-import logo from '../assets/djs_logo_rainbow_400x400.png';
-import vercelLogo from '../assets/powered-by-vercel.svg';
-import text from '../text.json';
+import { FiExternalLink } from 'react-icons/fi';
+import { PrismAsyncLight as SyntaxHighlighter } from 'react-syntax-highlighter';
+import { vscDarkPlus, ghcolors } from 'react-syntax-highlighter/dist/cjs/styles/prism';
 
-interface ButtonProps {
-	label: string;
-	href?: string;
-	ref?: Ref<HTMLAnchorElement>;
-	onClick?: MouseEventHandler<HTMLAnchorElement>;
-}
+const useStyles = createStyles((theme) => ({
+	outer: {
+		display: 'flex',
+		flexDirection: 'column',
+		height: '100%',
+		justifyContent: 'center',
+		gap: 50,
+		padding: 32,
 
-// eslint-disable-next-line react/display-name
-const LinkButton = forwardRef<HTMLAnchorElement, ButtonProps>(({ label, onClick, href }: ButtonProps, ref) => (
-	<a
-		href={href}
-		onClick={onClick}
-		ref={ref}
-		className="no-underline max-h-[70px] bg-blurple px-3 py-4 rounded-lg font-semibold text-white"
-	>
-		{label}
-	</a>
-));
+		[theme.fn.smallerThan('md')]: {
+			height: 'unset',
+		},
+	},
+
+	inner: {
+		display: 'flex',
+		justifyContent: 'space-between',
+		alignItems: 'center',
+
+		[theme.fn.smallerThan('md')]: {
+			flexDirection: 'column',
+			gap: 50,
+		},
+	},
+
+	content: {
+		maxWidth: 480,
+		marginRight: theme.spacing.xl,
+
+		[theme.fn.smallerThan('md')]: {
+			marginRight: 0,
+		},
+	},
+
+	title: {
+		color: theme.colorScheme === 'dark' ? theme.white : theme.black,
+		fontSize: 44,
+		lineHeight: 1.2,
+		fontWeight: 900,
+
+		[theme.fn.smallerThan('xs')]: {
+			fontSize: 28,
+		},
+	},
+
+	highlight: {
+		position: 'relative',
+		backgroundColor: theme.fn.variant({ variant: 'light', color: theme.primaryColor }).background,
+		borderRadius: theme.radius.sm,
+		padding: '4px 12px',
+	},
+}));
 
 export default function IndexRoute() {
+	const { classes } = useStyles();
+	const { colorScheme } = useMantineColorScheme();
+
 	return (
-		<main className="w-full max-w-full max-h-full h-full flex-col bg-white dark:bg-dark overflow-y-auto">
-			<div className="flex h-[65px] sticky top-0 border-b border-gray justify-center px-10 bg-white dark:bg-dark">
-				<div className="flex items-center w-full max-w-[1100px] justify-between">
-					<div className="h-[50px] w-[50px] rounded-lg overflow-hidden">
-						<Image className="h-[50px] w-[50px]" src={logo} />
-					</div>
-					<div className="flex flex-row space-x-8">
+		<Container className={classes.outer} size="lg">
+			<Box className={classes.inner}>
+				<Box className={classes.content}>
+					<Title className={classes.title}>
+						The <span className={classes.highlight}>most popular</span> way to build Discord <br /> bots.
+					</Title>
+					<Text color="dimmed" mt="md">
+						discord.js is a powerful Node.js module that allows you to interact with the Discord API very easily. It
+						takes a much more object-oriented approach than most other JS Discord libraries, making your bot&apos;s code
+						significantly tidier and easier to comprehend.
+					</Text>
+
+					<Group mt={30}>
 						<Link href="/docs" passHref>
-							<a className="no-underline text-blurple font-semibold">Docs</a>
+							<Button component="a" radius="sm" size="md">
+								Docs
+							</Button>
 						</Link>
-						<a className="text-blurple font-semibold">Guide</a>
-					</div>
-				</div>
-			</div>
-			<div className="xl:flex xl:flex-col xl:justify-center w-full max-w-full box-border p-10">
-				<div className="flex flex-col xl:flex-row grow max-w-[1100px] pb-10 space-y-10 xl:space-x-20 place-items-center place-self-center">
-					<div className="flex flex-col max-w-[800px] lt-xl:items-center">
-						<h1 className="font-bold text-6xl text-blurple my-2">{text.heroTitle}</h1>
-						<p className="text-xl text-dark-100 dark:text-gray-300">{text.heroDescription}</p>
-						<div className="flex flew-row space-x-4">
-							<LinkButton label="Read the guide" />
-							<Link href="/docs" passHref>
-								<LinkButton label="Check out the docs" />
-							</Link>
-						</div>
-					</div>
-					<div className="sm:flex sm:grow sm:shrink h-full xl:items-center hidden">
-						<Image src={codeSample} className="max-w-[600px] h-full rounded-xl shadow-md overflow-hidden" />
-					</div>
-				</div>
-				<div className="flex place-content-center">
-					<a href="https://vercel.com/?utm_source=discordjs&utm_campaign=oss" title="Vercel">
+						<Link href="https://discordjs.guide" passHref>
+							<Button component="a" variant="default" radius="sm" size="md" rightIcon={<FiExternalLink />}>
+								Guide
+							</Button>
+						</Link>
+					</Group>
+				</Box>
+				<Box pb="xs">
+					<SyntaxHighlighter
+						wrapLongLines
+						language="typescript"
+						style={colorScheme === 'dark' ? vscDarkPlus : ghcolors}
+						codeTagProps={{ style: { fontFamily: 'JetBrains Mono' } }}
+					>
+						{`import { Client, GatewayIntentBits } from 'discord.js';
+
+const client = new Client({ intents: [GatewayIntentBits.Guilds] });
+
+client.on('ready', () => {
+	console.log(\`Logged in as \${client.user.tag}!\`);
+});
+
+client.on('interactionCreate', async (interaction) => {
+	if (!interaction.isChatInputCommand()) return;
+
+	if (interaction.commandName === 'ping') {
+		await interaction.reply('Pong!');
+	}
+});
+
+await client.login('token');`}
+					</SyntaxHighlighter>
+				</Box>
+			</Box>
+			<Center>
+				<Link href="https://vercel.com/?utm_source=discordjs&utm_campaign=oss">
+					<a title="Vercel">
 						<Image
-							// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-							src={vercelLogo}
+							src="/powered-by-vercel.svg"
 							alt="Vercel"
-							className="max-w-[250px] shadow-md overflow-hidden"
+							width={0}
+							height={0}
+							style={{ height: '100%', width: '100%', maxWidth: 250 }}
 						/>
 					</a>
-				</div>
-			</div>
-		</main>
+				</Link>
+			</Center>
+		</Container>
 	);
 }
